+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata70.ts
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2024-2025 FlyByWire Simulations
 // SPDX-License-Identifier: GPL-3.0
 
-import { AbnormalProcedure } from 'instruments/src/MsfsAvionicsCommon/EcamMessages';
+import { AbnormalProcedure, ChecklistLineStyle } from 'instruments/src/MsfsAvionicsCommon/EcamMessages';
 
 // Convention for IDs:
 // First two digits: ATA chapter
@@ -986,13 +986,69 @@ export const EcamAbnormalSensedAta70: { [n: number]: AbnormalProcedure } = {
     items: [],
   },
   700900001: {
-    title: '\x1b<4m\x1b4mENG\x1bm RELIGHT IN FLIGHT (WIP)',
+    title: '\x1b<4m\x1b4mENG\x1bm RELIGHT IN FLIGHT',
     sensed: false,
-    items: [], // TODO
+    items: [
+      { name: 'FOR RELIGHT (SINGLE ENGINE)', sensed: false, condition: true },
+      { name: 'MAX GUARANTEED ALTITUDE : 28000 FT', sensed: false, level: 1 },
+      { name: 'MAX SPEED FOR WINDML RELIGHT : 250 KT', sensed: false, level: 1 },
+      { name: 'THR LEVER (AFFECTED)', sensed: false, labelNotCompleted: 'IDLE', level: 1 },
+      { name: 'ENG (AFFECTED) MASTER', sensed: false, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'XBLEED', sensed: true, labelNotCompleted: 'OPEN', level: 1 },
+      { name: 'WING A-ICE', sensed: true, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'FEED TK (AFFECTED ENG) MAIN PUMP', sensed: false, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'ENG START SEL', sensed: true, labelNotCompleted: 'IGN START', level: 1 },
+      { name: 'ENG (AFFECTED) MASTER', sensed: false, labelNotCompleted: 'ON', level: 1 },
+      { name: 'MAX EGT : 850 °C', sensed: false, level: 1 },
+      { name: 'ENG RELIGHT', sensed: false, labelNotCompleted: 'MONITOR (N3, FF, EGT)', level: 1 },
+      { name: 'WHEN IDLE REACHED (ENG AVAIL)', sensed: false, condition: true, level: 1 },
+      { name: 'ENG START SEL', sensed: true, labelNotCompleted: 'NORM', level: 2 },
+      { name: 'TCAS MODE', sensed: true, labelNotCompleted: 'TA/RA', level: 2 },
+      { name: 'XBLEED', sensed: true, labelNotCompleted: 'NORM', level: 2 },
+      { name: 'A-ICE', sensed: false, labelNotCompleted: 'AS RQRD', level: 2 },
+      { name: 'FOR RELIGHT (MULTIPLE ENGINES)', sensed: false, condition: true },
+      { name: 'THR LEVER (AFFECTED)', sensed: false, labelNotCompleted: 'IDLE', level: 1 },
+      { name: 'ENG START SEL', sensed: true, labelNotCompleted: 'IGN START', level: 1 },
+      { name: 'MAX GUARANTEED ALTITUDE : 28000 FT', sensed: false, level: 1 },
+      { name: 'MAX SPEED FOR WINDML RELIGHT : 250 KT', sensed: false, level: 1 },
+      { name: 'XBLEED', sensed: true, labelNotCompleted: 'OPEN', level: 1 },
+      { name: 'WING A-ICE', sensed: true, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'ENG (AFFECTED) MASTERS', sensed: false, labelNotCompleted: 'OFF 30S THEN ON', level: 1 },
+      { name: 'ENGs RELIGHT', sensed: false, labelNotCompleted: 'TRY REGULARLY', level: 1 },
+      { name: 'WHEN BELOW FL 200 : APU', sensed: false, labelNotCompleted: 'CONSIDER', level: 1 },
+      { name: 'WHEN BELOW FL 200', sensed: true, condition: true },
+      { name: 'APU', sensed: true, labelNotCompleted: 'START', level: 1 },
+      { name: 'WHEN APU AVAIL', sensed: true, condition: true, level: 1 },
+      { name: 'ENG (AFFECTED) MASTERS', sensed: false, labelNotCompleted: 'OFF', level: 2 },
+      { name: 'RECOM SPEED FOR START ASSIST : GDOT', sensed: false, style: ChecklistLineStyle.Green, level: 2 },
+      { name: 'APU BLEED', sensed: true, labelNotCompleted: 'ON', level: 2 },
+      { name: 'ENG (AFFCTD) MASTERS (2 AT A TIME)', sensed: false, labelNotCompleted: 'ON', level: 2 },
+      { name: 'ENG RELIGHT', sensed: false, labelNotCompleted: 'MONITOR (N3, FF, EGT)', level: 2 },
+      { name: 'ENGs RELIGHT TRY REGULARLY : ENGs (AFFCTD) OFF 30s THEN ON', sensed: false, level: 2 },
+      { name: 'IF NO FURTHER ENG RELIGHT ATTEMPT', sensed: false, condition: true },
+      { name: 'ENG (INOP) MASTER', sensed: false, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'ENG START SEL', sensed: true, labelNotCompleted: 'NORM', level: 1 },
+      { name: 'XBLEED', sensed: true, labelNotCompleted: 'NORM', level: 1 },
+      { name: 'A-ICE', sensed: false, labelNotCompleted: 'AS RQRD', level: 1 },
+      { name: 'IF NO ENG OUT', sensed: false, condition: true },
+      { name: 'TCAS MODE', sensed: true, labelNotCompleted: 'TA/RA', level: 1 },
+    ],
   },
   700900002: {
-    title: '\x1b<4m\x1b4mENG\x1bm TAIL PIPE FIRE (WIP)',
+    title: '\x1b<4m\x1b4mENG\x1bm TAIL PIPE FIRE',
     sensed: false,
-    items: [], // TODO
+    items: [
+      { name: 'ENG (AFFECTED) MASTER', sensed: false, labelNotCompleted: 'OFF' },
+      { name: 'XBLEED', sensed: true, labelNotCompleted: 'OPEN' },
+      { name: 'APU', sensed: true, labelNotCompleted: 'START' },
+      { name: 'APU BLEED', sensed: true, labelNotCompleted: 'ON' },
+      { name: 'WHEN N3 BELOW 30%', sensed: false, condition: true },
+      { name: 'ENG START SEL', sensed: true, labelNotCompleted: 'CRANK', level: 1 },
+      { name: 'ENG (AFFECTED) MAN START', sensed: false, labelNotCompleted: 'ON', level: 1 },
+      { name: 'WHEN FIRE STOPPED', sensed: false, condition: true },
+      { name: 'ENG (AFFECTED) MAN START', sensed: false, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'ENG START SEL', sensed: true, labelNotCompleted: 'NORM', level: 1 },
+      { name: 'XBLEED', sensed: true, labelNotCompleted: 'AUTO', level: 1 },
+    ],
   },
 };