+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/AbnormalSensed/ata80-rest.ts
@@ -19,9 +19,37 @@ import { AbnormalProcedure, ChecklistLineStyle } from 'instruments/src/MsfsAvion
 /** All abnormal sensed procedures (alerts, via ECL) should be here. */
 export const EcamAbnormalSensedAta80Rest: { [n: number]: AbnormalProcedure } = {
   990900001: {
-    title: '\x1b<4m\x1b4mMISC\x1bm BOMB ON BOARD (WIP)',
+    title: '\x1b<4m\x1b4mMISC\x1bm BOMB ON BOARD',
     sensed: false,
-    items: [], // TODO
+    items: [
+      { name: 'CKPT / CABIN COM', sensed: false, labelNotCompleted: 'ESTABLISH' },
+      { name: 'IF LDG + EVAC POSSIBLE WITHIN 30 MIN', sensed: false, condition: true },
+      { name: 'ATC', sensed: false, labelNotCompleted: 'NOTIFY', labelCompleted: 'NOTIFIED', level: 1 },
+      { name: 'COMPANY', sensed: false, labelNotCompleted: 'NOTIFY', labelCompleted: 'NOTIFIED', level: 1 },
+      { name: 'EVAC', sensed: false, labelNotCompleted: 'PREPARE', level: 1 },
+      { name: 'IF NO IMMEDIATE LANDING', sensed: false, condition: true },
+      { name: 'ATC', sensed: false, labelNotCompleted: 'NOTIFY', labelCompleted: 'NOTIFIED', level: 1 },
+      { name: 'COMPANY', sensed: false, labelNotCompleted: 'NOTIFY', labelCompleted: 'NOTIFIED', level: 1 },
+      { name: 'ACFT (IF CLIMBING)', sensed: false, labelNotCompleted: 'LEVEL OFF', level: 1 },
+      { name: 'TRGT SPEED : PREFER LO IAS', sensed: false, level: 1 },
+      { name: 'CABIN ALT MODE', sensed: true, labelNotCompleted: 'MAN', level: 1 },
+      { name: 'DESCENT TO CAB ALT + 2500FT/MEA-MORA', sensed: false, level: 1 },
+      { name: 'WHEN ACFT ALT = CAB ALT + 2500FT/MEA-MORA', sensed: false, condition: true, level: 2 },
+      { name: 'ELEC GALLEY', sensed: true, labelNotCompleted: 'OFF', level: 2 },
+      { name: 'ELEC PAX SYS', sensed: true, labelNotCompleted: 'OFF', level: 2 },
+      { name: 'IF FUEL PERMITS', sensed: false, condition: true, level: 2 },
+      { name: 'FLAPS', sensed: true, labelNotCompleted: 'AT LEAST CONF 1', level: 2 },
+      { name: 'L/G LEVER', sensed: true, labelNotCompleted: 'DOWN', level: 2 },
+      { name: 'FUEL CONSUMPT INCRSD', sensed: false, level: 1 },
+      {
+        name: 'FMS PRED UNRELIABLE WITHOUT ACCURATE FMS FUEL PENALTY INSERTION',
+        sensed: false,
+        style: ChecklistLineStyle.Green,
+        level: 1,
+      },
+      { name: 'DURING FURTHER DESCENT : MAINTAIN MAX DIFF PRESS 1 PSI', sensed: false, level: 1 },
+      { name: 'EVAC', sensed: false, labelNotCompleted: 'PREPARE', level: 1 },
+    ],
   },
   990900002: {
     title: '\x1b<4m\x1b4mMISC\x1bm CKPT WINDOW CRACKED (WIP)',
@@ -122,13 +150,38 @@ export const EcamAbnormalSensedAta80Rest: { [n: number]: AbnormalProcedure } = {
     ],
   },
   990900010: {
-    title: '\x1b<4m\x1b4mMISC\x1bm SEVERE TURBULENCE (WIP)',
+    title: '\x1b<4m\x1b4mMISC\x1bm SEVERE TURBULENCE',
     sensed: false,
-    items: [], // TODO
+    items: [
+      { name: 'SEAT BELTS', sensed: true, labelNotCompleted: 'ON' },
+      { name: 'MAX TURB SPEED : 300 / .85 ', sensed: false },
+      { name: 'MIN TURB SPEED : GREEN DOT', sensed: false },
+      { name: 'ADJUST SPEED AS NECCESSARY FOR COMFORT', sensed: false },
+      { name: 'AP : KEEP ON', sensed: false },
+      { name: 'CABIN & CKPT (LOOSE EQPT)', sensed: false, labelNotCompleted: 'SECURE' },
+      { name: 'SPEED BRAKES', sensed: false, labelNotCompleted: 'AS RQRD' },
+      { name: 'IF EXCESSIVE THRUST VARIATIONS', sensed: false, condition: true },
+      { name: 'A/THR', sensed: true, labelNotCompleted: 'OFF', level: 1 },
+      { name: 'DESCENT TO OR BELOW OPT ALT', sensed: false, labelNotCompleted: 'CONSIDER' },
+    ],
   },
   990900011: {
-    title: '\x1b<4m\x1b4mMISC\x1bm VOLCANIC ASH ENCOUNTER (WIP)',
+    title: '\x1b<4m\x1b4mMISC\x1bm VOLCANIC ASH ENCOUNTER',
     sensed: false,
-    items: [], // TODO
+    items: [
+      { name: '180° TURN', sensed: false, labelNotCompleted: 'INITIATE' },
+      { name: 'ATC', sensed: false, labelNotCompleted: 'NOTIFY' },
+      { name: 'A/THR', sensed: true, labelNotCompleted: 'OFF' },
+      { name: 'THRUST (IF CONDS PERMIT)', sensed: false, labelNotCompleted: 'REDUCE' },
+      { name: 'CREW OXY MASKS', sensed: false, labelNotCompleted: 'USE / 100% / EMER' },
+      { name: 'CABIN CREW', sensed: false, labelNotCompleted: 'NOTIFY' },
+      { name: 'PAX OXY MASK MAN ON', sensed: false, labelNotCompleted: 'AS RQRD' },
+      { name: 'ENGs A-ICE', sensed: true, labelNotCompleted: 'ON' },
+      { name: 'WING A-ICE', sensed: true, labelNotCompleted: 'ON' },
+      { name: 'AIR FLOW', sensed: true, labelNotCompleted: 'HI' },
+      { name: 'ENG PARAMETERS', sensed: false, labelNotCompleted: 'MONITOR' },
+      { name: 'AIR SPEED', sensed: false, labelNotCompleted: 'MONITOR' },
+    ],
   },
 };
+// FIXME: add defered proc for BOMB ON BOARD