+++ fbw-a380x/src/systems/instruments/src/MsfsAvionicsCommon/EcamMessages/index.ts
@@ -252,6 +252,10 @@ export const EcamLimitations: { [n: string]: string } = {
   800400002: '\x1b<5mLDG DIST AFFECTED',
   800400003: '\x1b<5mLDG PERF AFFECTED',
   800400004: '\x1b<5mFOR GA : KEEP S/F CONF',
+  800400005: '\x1b<5mMAX DIFF PR : 1 PSI',
+  800400006: '\x1b<5mMAX SPEED : 300 / .85',
+  800400007: '\x1b<5mMIN SPEED : GREEN DOT',
+  800400008: '\x1b<5m\x1b5mTURB SPEED:',
 };
 
 /** All possible INOP sys, with special formatting characters. */